refactor(cli): split commands/mod.rs by subcommand (#1126)
Behavior-preserving module split of the 2,997-line commands/mod.rs into
per-concern sibling modules under commands/, leaving mod.rs as a thin
dispatcher (module decls + pub use re-exports) so callers keep using
crate::commands::<fn> unchanged.

New modules (verbatim relocation):
- support.rs        stub parser/result types + resolve_sstable_path + display_*
- schema_load.rs    load_schema_file + parse_json_schema (+ dropped-column tests)
- query.rs          execute_query (+cfg) + execute_select_query + print_csv_format
- import.rs         import_data + CSV/JSON batch import helpers
- export.rs         export_data (streaming) + export_to_* helpers
- read.rs           read_sstable + read_sstable_enhanced
- export_sstable.rs export_sstable + export_as_* + parquet helpers
- inspect.rs        validate_sstable + analyze_sstable
- benchmark_sstable.rs benchmark_sstable + benchmark_*_operation

No semantic changes. mod.rs: 2,997 -> 51 lines; each new module <=671 lines.
Relocated the 3 dropped_column_json tests with their code (count preserved).

Pre-existing failures unchanged vs origin/main (verified): the 3
cli_schema_validation DML tests, test_duckdb_schema_inference,
test_corrupted_file_handling, test_collections_table_output_snapshot;
status_metrics::test_is_stale is timing-flaky (passes on rerun).
